Revert "[Audio] Connect to lavalink in the background (#2335)"
This reverts commit b633a33137d11404e34c6d3eee4bbc3f6061f291.
